korjaa status

+++ backend/src/sqsEvents/sqsEventHandlerLambda.ts
@@ -31,6 +31,10 @@ import {
 } from "../util/lausuntoPyyntoUtil";
 import { velho } from "../velho/velhoClient";
 import { linkHyvaksymisPaatos } from "hassu-common/links";
+import { deleteFilesUnderSpecifiedVaihe } from "../HyvaksymisEsitys/s3Calls/deleteFiles";
+import { ENNAKKONEUVOTTELU_JULKAISU_PATH, ENNAKKONEUVOTTELU_PATH } from "../ennakkoneuvottelu/tallenna";
+import { getHyvaksymisEsityksenLadatutTiedostot } from "../HyvaksymisEsitys/getLadatutTiedostot";
+import getHyvaksymisEsityksenAineistot from "../HyvaksymisEsitys/getAineistot";
 
 async function handleNahtavillaoloZipping(ctx: ImportContext) {
   if (!ctx.projekti.nahtavillaoloVaihe) {
@@ -481,7 +485,24 @@ export const handlerFactory = (event: SQSEvent) => async () => {
 async function checkEpaaktiivinen(ctx: ImportContext) {
   if (ctx.projektiStatus == API.Status.EPAAKTIIVINEN_1) {
     await aineistoDeleterService.deleteAineistoIfEpaaktiivinen(ctx);
-    await projektiDatabase.deleteEnnakkoNeuvottelu(ctx.oid);
+    if (ctx.projekti.ennakkoNeuvottelu || ctx.projekti.ennakkoNeuvotteluJulkaisu) {
+      await projektiDatabase.deleteEnnakkoNeuvottelu(ctx.oid);
+    }
+    if (ctx.projekti.ennakkoNeuvotteluJulkaisu) {
+      const tiedostot = await getHyvaksymisEsityksenLadatutTiedostot(ctx.projekti.ennakkoNeuvotteluJulkaisu);
+      const aineistot = await getHyvaksymisEsityksenAineistot(ctx.projekti.ennakkoNeuvotteluJulkaisu);
+      await deleteFilesUnderSpecifiedVaihe(
+        ctx.oid,
+        ENNAKKONEUVOTTELU_JULKAISU_PATH,
+        [...tiedostot, ...aineistot, { avain: ENNAKKONEUVOTTELU_JULKAISU_PATH + "/aineisto.zip", nimi: "aineisto.zip" }],
+        "Projekti epäaktiivinen"
+      );
+    }
+    if (ctx.projekti.ennakkoNeuvottelu) {
+      const tiedostot = await getHyvaksymisEsityksenLadatutTiedostot(ctx.projekti.ennakkoNeuvottelu);
+      const aineistot = await getHyvaksymisEsityksenAineistot(ctx.projekti.ennakkoNeuvottelu);
+      await deleteFilesUnderSpecifiedVaihe(ctx.oid, ENNAKKONEUVOTTELU_PATH, [...tiedostot, ...aineistot], "Projekti epäaktiivinen");
+    }
   }
 }
